refactor: relocate auth-only entry
