+++ README.md
@@ -79,14 +79,17 @@ that required exports exist and that no compiled file still contains
 The storefront SDK publishes lightweight bundles that can be loaded directly:
 
 ```html
-<!-- Auth-only bundle -->
-<script type="module" src="https://sdk.smoothr.io/core/auth-only.js"></script>
+<!-- SDK auth entry bundle -->
+<script
+  type="module"
+  src="https://sdk.smoothr.io/features/auth/sdk-auth-entry.js"
+></script>
 
 <!-- Checkout bundle -->
 <script type="module" src="https://sdk.smoothr.io/core/checkout.js"></script>
 ```
 
-`auth-only.js` initializes authentication and currency helpers without any
+`sdk-auth-entry.js` initializes authentication and currency helpers without any
 checkout logic. `checkout.js` provides the checkout flow and can be paired with
 platform adapters like `./platforms/webflow/checkout.js`.
