+++ storefronts/adapters/core/sdk-auth-entry.js
@@ -1,0 +1,2 @@
+export * from '../../features/auth/sdk-auth-entry.js';
+export { default } from '../../features/auth/sdk-auth-entry.js';
