--- storefronts/platforms/core/auth-only.js
@@ -1,2 +1,0 @@
-export * from '../../core/auth-only.js';
-export { default } from '../../core/auth-only.js';
